Resolve Some of Logan's Changes
