+++ requirements.txt
@@ -6,12 +6,9 @@ six>=1.11.0
 h5py>=2.10.0
 numpy>=1.15.4
 pandas>=0.23.4
-scikit-learn>=1.0
 pydantic>=1.6.1
 joblib>=1.0.0
 mdf_forge>=0.8.0
 mdf-connect-client>=0.4.0
 json2table>=1.1.5
 joblib>=1.1.0
-torch>=1.8.0
-tensorflow>=2.8
